# difflib: optimize SplitLines