Use strings.SplitAfter to avoid an allocation and copy per line.
    
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkSplitLines100       25744         7168          -72.16%
BenchmarkSplitLines10000     2227947       369944        -83.40%

benchmark                    old allocs     new allocs     delta
BenchmarkSplitLines100       109            1              -99.08%
BenchmarkSplitLines10000     10021          1              -99.99%

benchmark                    old bytes     new bytes     delta
BenchmarkSplitLines100       6144          1664          -72.92%
BenchmarkSplitLines10000     1030960       163840        -84.11%
```